requirements.txt
Django==2.2.15
django-q==1.3.2
djangorestframework==3.12.4
eventlet==0.32.0
Faker==9.8.0
gevent==21.8.0
redis==3.5.3
django-cors-headers==3.10.0
psycopg2-binary
celery[redis]
django-celery-results
flower
